ml-tasks/task-3/requirements.txt
streamlit
langchain>=0.0.300
chromadb
sentence-transformers
pypdf
python-dotenv
google-generativeai 